# Add expirable LRU implementation, with generics