My attempt at expirable cache implementation is similar to what was done in #41. I've tried to make the new cache as close as possible to the existing one regarding code style and behaviour.

The only difference I left in place is that `size=0` on this type of cache means _unlimited_, which makes sense with the expirable cache.

Original work was done in [lcw](https://github.com/paskal/lcw/tree/lru_internal), but it turned out that I don't need it there, and I thought it would be nice to add my changes to upstream (this repo).

This PR is a recreation of #68, which was not merged. cc @jefferai @Dentrax for review.